working on DicomStreamReader

+++ OrthancFramework/Sources/DicomFormat/DicomMap.cpp
@@ -506,10 +506,10 @@ namespace Orthanc
   {
     const uint8_t* p = reinterpret_cast<const uint8_t*>(dicom);
 
-    return (static_cast<uint16_t>(p[0]) |
-            (static_cast<uint16_t>(p[1]) << 8) |
-            (static_cast<uint16_t>(p[2]) << 16) |
-            (static_cast<uint16_t>(p[3]) << 24));
+    return (static_cast<uint32_t>(p[0]) |
+            (static_cast<uint32_t>(p[1]) << 8) |
+            (static_cast<uint32_t>(p[2]) << 16) |
+            (static_cast<uint32_t>(p[3]) << 24));
   }
